config jwt security

+++ src/main/java/com/example/entity/UserVocabulary.java
@@ -39,7 +39,7 @@ public class UserVocabulary {
   private Date endDate;
 
   private short countFalse;
-  private int timeRepetition;
+  private long timeRepetition;
 
   public UserVocabulary(User user, Vocabulary vocabulary) {
     this.id = new UserVocabularyId(user.getId(), vocabulary.getId());
